+++ src/map_props.c
@@ -6,7 +6,7 @@
 /*   By: iostancu <iostancu@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/05/31 17:21:37 by iostancu          #+#    #+#             */
-/*   Updated: 2022/06/16 20:23:24 by iostancu         ###   ########.fr       */
+/*   Updated: 2022/06/21 23:55:22 by iostancu         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -29,8 +29,8 @@ void	set_new_zoom(t_fdf *fdf)
 
 	segm.x = fdf->mtrx.segments.x;
 	segm.y = fdf->mtrx.segments.y;
-	fdf->mtrx.px_size.x = (int)((segm.x - 1) * fdf->mtrx.zoom);
-	fdf->mtrx.px_size.y = (int)((segm.y - 1) * fdf->mtrx.zoom);
+	fdf->mtrx.px_size.x = ((segm.x - 1) * fdf->mtrx.zoom);
+	fdf->mtrx.px_size.y = ((segm.y - 1) * fdf->mtrx.zoom);
 	set_new_limits(fdf);
 }
 
@@ -50,6 +50,10 @@ void	set_new_limits(t_fdf *fdf)
 	fdf->mtrx.end_draw.y = fdf->mtrx.start_draw.y + fdf->mtrx.px_size.y;
 }
 
+// void	set_limits_isometric(t_fdf *fdf)
+// {
+	
+// }
 // t_img_props	get_middle_img()
 // {
 	// t_vec2	save_aux;
